+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Release Date: ?
  - [BC] renamed Peak\Bedrock\Controller\Action to Peak\Bedrock\Controller\ActionController
  - added an exception in Peak\Bedrock\View\Form\FormBuilder::control()
  - added Peak\Common\TextUtils move code from Peak\Bedrock\View\Helper\Text there
+ - added function formatFileSize() to Peak\Common\helpers.php
+ - deleted deprecated class Peak\Common\Spl\Fileinfo and Peak\Common\Spl\Dirinfo
 
 VERSION 2.10.0
 --------------